refactor(health): split HealthDiagnosisPage into focused sub-components
